Verbose mode implementation
Added a 'verbose' mode for easier debugging, which prints the modified values of the propositions and connectives while the algorithm is running.

Added a 'determine(...)' function, which gets passed a variable and determines its value.

Added a 'summary(...)' function, which takes any number of propositions and/or connectives and represents them in an organised visual table.