Major changes
Removed annotations at the beginning of the script.

Added documentation for all classes and functions. Properly structured it, as well as some methods inside the classes.

Added square brackets as enclosing method for visual differentiation of connectives and propositions.

Added missing self.check() calls in Implicative and BiImplicative classes.

Fixed wrong inheritance of the BiImplicative connective from And to BinaryConnective.

Added a test for Or and XOr connectives.

Module documentation pending.